Migrate to AGP 9.0.0 and the new per-platform KMP structure
Update the embedded submodules to latest main (Enro 3.0.0-beta02, Udytils
AGP 9.0.0 migration) and bring the template up to date.

- AGP 8.13.2 -> 9.0.0. KMP modules now use com.android.kotlin.multiplatform.library
  with the `kotlin { androidLibrary { } }` accessor (pinned at 9.0.0 for IntelliJ
  support; `android {}` alias is 9.1+). Gradle wrapper 8.14.3 -> 9.5.0.
- Split :app:client into per-platform modules per the new KMP default structure:
  :app:client:{shared,android,desktop,web}. `shared` is a Compose KMP library
  (+ iOS App.framework); android/desktop/web are thin per-platform applications.
- Remove kotlinx-rpc (unused scaffolding; no stable release supports Kotlin
  2.3.21). Plain Ktor + kotlinx-serialization remains the client/server transport.
- Bump versions: Kotlin 2.3.0 -> 2.3.21, KSP -> 2.3.9, Ktor -> 3.5.0, Koin -> 4.2.1,
  serialization -> 1.11.0, datetime -> 0.8.0, io -> 0.9.0, androidx-activity/core,
  logback. Compose 1.10.3 / lifecycle 2.9.6 / coroutines 1.10.2 pinned to match the
  submodules' merged runtime.
- Drop iosX64 (removed upstream from the submodules).
- Fix a stale generated-accessor import in build-logic exposed by the Gradle bump.
