+++ app/client/android/src/main/AndroidManifest.xml
