+++ app/client/desktop/build.gradle.kts
@@ -1,0 +1,54 @@
+import org.jetbrains.compose.desktop.application.dsl.TargetFormat
+import org.jetbrains.kotlin.gradle.dsl.JvmTarget
+import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
+
+plugins {
+    alias(libs.plugins.kotlinJvm)
+    alias(libs.plugins.composeMultiplatform)
+    alias(libs.plugins.composeCompiler)
+}
+
+private val projectNamespace = providers.gradleProperty("ukpt.projectNamespace").get()
+
+// Keep Java compilation aligned with the Kotlin jvmTarget (11) to satisfy the
+// JVM-target consistency check on this plain Kotlin/JVM module.
+java {
+    sourceCompatibility = JavaVersion.VERSION_11
+    targetCompatibility = JavaVersion.VERSION_11
+}
+
+dependencies {
+    implementation(projects.app.client.shared)
+
+    implementation(compose.desktop.currentOs)
+    implementation(libs.kotlinx.coroutinesSwing)
+}
+
+compose.desktop {
+    application {
+        mainClass = "$projectNamespace.MainKt"
+
+        nativeDistributions {
+            targetFormats(
+                TargetFormat.Dmg,
+                TargetFormat.Msi,
+                TargetFormat.Deb,
+            )
+            packageName = projectNamespace
+            packageVersion = "1.0.0"
+        }
+    }
+}
+
+// Preview-feature flags match the KMP convention plugins so this module can consume
+// :app:client:shared, whose output is marked pre-release by those features.
+tasks.withType<KotlinCompile>().configureEach {
+    compilerOptions {
+        jvmTarget.set(JvmTarget.JVM_11)
+        freeCompilerArgs.addAll(
+            "-Xcontext-parameters",
+            "-Xexplicit-backing-fields",
+            "-Xexpect-actual-classes",
+        )
+    }
+}
