+++ app/client/shared/src/iosSimulatorArm64Main/kotlin/com/isaacudy/ukpt/installUkptNavigation.kt
