--- app/client/src/iosX64Main/kotlin/com/isaacudy/ukpt/installUkptNavigation.kt
@@ -1,7 +1,0 @@
-package com.isaacudy.ukpt
-
-import platform.UIKit.UIApplication
-
-internal actual fun installUkptNavigation() {
-    UkptNavigation.installNavigationController(UIApplication.sharedApplication())
-}
