+++ app/server/build.gradle.kts
@@ -2,7 +2,6 @@ plugins {
     id("ukpt.jvm-server")
     alias(libs.plugins.ktor)
     alias(libs.plugins.kotlinSerialization)
-    alias(libs.plugins.kotlinxRpc)
 }
 
 private val projectNamespace = providers.gradleProperty("ukpt.projectNamespace").get()
@@ -26,10 +25,6 @@ dependencies {
     implementation(libs.ktor.serverNetty)
     implementation(libs.ktor.serverWebsockets)
     implementation(libs.ktor.server.auth)
-
-    implementation(libs.kotlinx.rpc.krpcServer)
-    implementation(libs.kotlinx.rpc.krpcKtorServer)
-    implementation(libs.kotlinx.rpc.krpcSerializationJson)
 
     testImplementation(libs.ktor.serverTestHost)
     testImplementation(libs.kotlin.testJunit)
