+++ feature/core/api/build.gradle.kts
@@ -1,27 +1,20 @@
 plugins {
     id("ukpt.kmp-library")
     alias(libs.plugins.kotlinSerialization)
-    alias(libs.plugins.kotlinxRpc)
-}
-
-android {
-    namespace = "feature.core.api"
 }
 
 kotlin {
+    @Suppress("UnstableApiUsage")
+    androidLibrary {
+        namespace = "feature.core.api"
+    }
     sourceSets {
         commonMain.dependencies {
-            implementation(libs.kotlinx.rpc.krpcSerializationJson)
-            implementation(libs.kotlinx.datetime)
             implementation(libs.ktor.websockets)
 
-            compileOnly(libs.kotlinx.rpc.krpcClient)
-            compileOnly(libs.kotlinx.rpc.krpcServer)
             api(libs.kotlinx.coroutinesCore)
             api(libs.kotlinx.serialization)
             api(libs.kotlinx.datetime)
-            api(libs.kotlinx.rpc.core)
-            api(libs.kotlinx.rpc.krpcCore)
             api(libs.enro.common)
             api(libs.udytils.core)
         }
