+++ feature/core/server/build.gradle.kts
@@ -1,7 +1,6 @@
 plugins {
     id("ukpt.jvm-library")
     alias(libs.plugins.kotlinSerialization)
-    alias(libs.plugins.kotlinxRpc)
 }
 
 dependencies {
@@ -14,9 +13,6 @@ dependencies {
     implementation(libs.ktor.clientContentNegotiation)
     implementation(libs.ktor.serialization.kotlinx.json)
 
-    implementation(libs.kotlinx.rpc.krpcServer)
-    implementation(libs.kotlinx.rpc.krpcKtorServer)
-    implementation(libs.kotlinx.rpc.krpcSerializationJson)
     implementation(libs.enro.common)
     implementation(libs.udytils.core)
     implementation(libs.koin.core)
